+++ dwds/lib/src/debugging/dart_runtime_debugger.dart
@@ -18,11 +18,8 @@ class DartRuntimeDebugger {
   String _generateJsExpression(
     String ddcExpression,
     String libraryBundleExpression,
-  ) {
-    return _useLibraryBundleExpression
-        ? libraryBundleExpression
-        : ddcExpression;
-  }
+  ) =>
+      _useLibraryBundleExpression ? libraryBundleExpression : ddcExpression;
 
   /// Wraps a JS function call with SDK loader logic.
   String _wrapWithSdkLoader(String args, String functionCall) {
@@ -200,14 +197,38 @@ class DartRuntimeDebugger {
     );
   }
 
-  String getVariableJsExpression(
-    String libraryName,
-    String url,
-    String variable,
-  ) {
+  /// Generates a JS expression to invoke a Dart extension method.
+  String invokeExtensionJsExpression(String methodName, String encodedJson) {
     return _generateJsExpression(
-      "${_loadStrategy.loadModuleSnippet}('dart_sdk').dart.getModuleLibraries('$libraryName')['$url']['$variable'];",
-      'dartDevEmbedder.dart.getModuleLibraries("$libraryName")["$url"]["$variable"]',
+      "${_loadStrategy.loadModuleSnippet}('dart_sdk').developer.invokeExtension('$methodName', JSON.stringify($encodedJson));",
+      "dartDevEmbedder.debugger.invokeExtension('$methodName', JSON.stringify($encodedJson));",
+    );
+  }
+
+  /// Generates a JS expression for calling a library method.
+  String callLibraryMethodJsExpression(
+    String libraryUri,
+    String methodName,
+  ) {
+    final findLibraryExpression = '''
+     (function() {
+       const sdk = ${_loadStrategy.loadModuleSnippet}('dart_sdk');
+       const dart = sdk.dart;
+       const library = dart.getLibrary('$libraryUri');
+       if (!library) throw 'cannot find library for $libraryUri';
+       return library;
+     })();
+     ''';
+
+    // `callLibraryMethod` expects an array of arguments. Chrome DevTools spreads
+    // arguments individually when calling functions. This code reconstructs the
+    // expected argument array.
+    return _generateJsExpression(
+      findLibraryExpression,
+      _wrapWithBundleLoader(
+        '',
+        'callLibraryMethod("$libraryUri", "$methodName", Array.from(arguments))',
+      ),
     );
   }
 }